requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
contourpy==1.3.1
crispy-bootstrap4==2024.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
Django==5.1
django-crispy-forms==2.3
django-filter==24.3
docstring-to-markdown==0.15
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jedi==0.19.2
jsonpointer==2.1
kiwisolver==1.4.8
matplotlib==3.9.2
multidict==6.1.0
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pyls-spyder==0.4.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.16
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
requests==2.32.3
selenium==4.27.1
service-identity==24.2.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
ujson==5.10.0
unicode-slugify==0.1.5
Unidecode==1.3.8
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
